requirements.txt
Flask==3.1.2
PyMuPDF==1.23.8
python-docx==1.1.0
gunicorn==21.2.0
